Move the logic for killing into set_inputs
In thinking about how to handle networking, state is expected to get out of sync sometimes. That's OK for stuff like position, but we can't have life and death get out of sync! And since who is killed is based on position, that could happen!

So, simulate handles inputs for all players as far as positioning goes, but we'll send a special message to indicate when a player dies (not yet implemented)