Initial kinda working sync over the network!
So, the borrow checker was indeed unhappy with there being two
flows where changes could come in – the network and the main JS loop.

The solution was to use a wrapper to synchronize, an Arc<Mutex<Game>>.
I think this is a fairly common pattern for dealing with mutation
from multiple sources.

Arc<X> is a reference counted wrapper around X. It's like Box<X> only
there can be more than one copy of the pointer to the value, and
the data will stick around as long as at least one pointer lives. Docs here: https://doc.rust-lang.org/std/sync/struct.Arc.html

Mutex<X> defines a read/write lock around X, such that you can only
read or write the underlying X value while you have a lock on it.
When you're done with the value you automatically release the lock.
Pretty handy! I've always seen it used with Arc, I guess because if
there's only one copy of the Mutex then there's no point in the Mutex
because ordinary Rust lifetimes give you the guarantees of a lock.
Docs here: https://doc.rust-lang.org/std/sync/struct.Mutex.html

I _think_ that we'll never be in a situation where the lock is held
by multiple things at the same time, because of the guarantees of the
browser's event loop – we're fundamentally single threaded at this
point. However Rust can't know that, and in the future we can use
SharedArrayBuffer anyways, so this seems like good discipline.

Also I don't know how to opt out anyways, so it's not like we have any
other options lol.

So with this in mind, we don't pass back a real Game to the JS, just
a wrapper around an Arc<Mutex<Game>>.

In a follow up change I'll clean up the simulate / simulate_internal now
that Game is internal.