Add a join step
Players are now identified by a UUID, and when they try to join they propose a color. Other players allow this if and only if that color isn't taken. Otherwise they get the first available color.

Snapshot might as well be named JoinResponse at this point.

Note that one Join message will be followed by N-1 Snapshots, where N is the number of players.

Note that we assume that messages are not reordered. I _believe_ this is a valid expectation.