Abstract out the websocket field on Game
With this change, the socket that game has can only be used for sending messages.

Changes to initialization are then necessary, because we can't use the socket on Game to access the websocket directly, we need to initialize the websocket before we even have a game object, meaning that we need a super awkward mutex around an optional mutex...

Seems to work all right though, and it's a positive step in two directions: initializing the websocket connection before returning Game, and being able to test the core game logic from a pure rust test without going through the browser.