Move canonical logic onto the server.
This is a big change, but I think it will serve us well going forward.

This separates out three types: Game, GameAsPlayer, and GameServer.

Game is the core game state: settings like kill distance, players and
their complete state (impostor, tasks, etc), dead bodies, etc, as well
as all the info needed to extrapolate this state into the future
assuming no new user input (e.g. everyone's velocities).

GameServer is a Game plus a Broadcaster which is able to communicate
with players. The server in server/src/main.rs implements a Broadcaster
and calls into the appropriate methods on GameServer for events like
client connection, message received, and client disconnection.

GameAsPlayer is a Game plus an interface for handling user input and
a GameTx for communicating those inputs up to the server.
Currently GameAsPlayer does validation of actions (like, "am I allowed
to finish this task") but in a future change that will move up to the
server.