Refine the lifecycle of a game
The server is initially doing nothing. When the first player connects, we start simulating the game every 16ms until the game finishes (one side wins or every player diconnects).

When a player joins and the current game is finished, instead create a new game.

Also we now apportion out tasks and impostor status from the server, not from the player that starts the game.