Add a transition from night to day.
Players have a certain amount of time to vote. At the end of that time, or once all eligable players have voted, we tally up the votes and kill the lucky winner! The crew can also vote to skip killing for some unknowable reason.

Also added some more logic to player disconnection. Initially this was just giving players who voted for the disconnected player their votes back, but I also added checks to see if one of the teams won.