Refine the debug replay experience.
To replay the most recent game that completed, go to
http://localhost:8080/?recording

Like YouTube, J jumps back 5s, L jumps forward 5s, and K pauses/unpauses.

This fixes a bunch of subtle bugs in playback, getting it pretty solid.

Also when you replay, it will attempt to force playback of the recording
in the current version, even though that's incoherent. If you don't
change core/ though, there's a good chance it'll work.