Improve incremental builds by 710% by dropping webpack
I spent this morning messing with drawing a voting screen, which for someone who isn't good at geometry involves a lot of trial and error, and the incremental build perf started to really get to me.

So, currently we're using webpack to bundle our JS for production, and for a devserver. That came in from the wasm-pack starter kit so I figured it was a reasonable enough choice. However, we've got just a couple KB of JS, basically the entire project is written in Rust, and it seemed like webpack was taking a decent amount of time to build, so I wanted to see if we could do better.

I started out doing a few timing runs of incremental builds and wasm-pack -> webpack commands were together taking ~12s. The first thing I tried was to upgrade to webpack 5, which is supposed to have better caching and maybe be faster. Looks like it makes our build about 1s slower, so that's no help.

Looking at the output that wasm-pack was producing, it included a synchronous import of the generated `.wasm` file which isn't yet standard (though [import assertions](https://github.com/tc39/proposal-import-assertions) are probably going to unblock making that work, the syntax is likely to be `import game from "./client.wasm" assert { type: "wasm" };`).

It turns out that wasm-pack by default runs with [`--target bundler`](https://rustwasm.github.io/wasm-pack/book/commands/build.html#target) which emits code designed for webpack to consume. Kinda gross to emit pseudostandard output by default, but it means that your imports can all by synchronous which is kinda nice. But there's a `--target web` version, [what's that do](https://rustwasm.github.io/docs/wasm-bindgen/reference/deployment.html#without-a-bundler)?

Well, while the import syntax is worked out, there are some interim methods for loading and instantiating a WebAssembly module. Because they involve loading data from the network they're asynchronous, but that's no big deal, our initialization function is already async, so we can just add an `async initWasm()` no problem. With that, we can drop webpack entirely, bringing us down from 12s to 4s.

**3x faster**!? Our 90% Rust project that has one JS file and zero JS dependencies was spending 3s in webpack for every second in the Rust compiler? The infamously slow Rust compiler that everyone warns you about when you say you're gonna try out this new Rust thing?

But then, poking around a bit, it turns out that's just the start! Because I switched us run `wasm-pack` with the `--release` flag even during development because it was net faster, as the larger unoptimized wasm file made webpack take tens of seconds to process. With webpack out of the picture we can afford to run in dev mode, which gets the build time down to `1.69s`! That's **7x faster**!

The difference is night and day, finally I can quickly iterate again!

Now potentially we'd still want to do some bundling, maybe if we add more JS to our repo. Seems like esbuild is a good option there if we go that route. For a devserver I'm using `@web/dev-server` just for livereload, but it can also downlevel if we end up needing that.